+++ src/main/java/com/nylas/FreeBusy.java
@@ -5,6 +5,7 @@ import java.util.List;
 public class FreeBusy {
 
 	private String email;
+	private String calendar_id;
 	private List<TimeSlot> time_slots;
 	
 	public String getEmail() {
@@ -13,6 +14,10 @@ public class FreeBusy {
 
 	public List<TimeSlot> getTimeSlots() {
 		return time_slots;
+	}
+
+	public String getCalendarId() {
+		return calendar_id;
 	}
 
 	public void setEmail(String email) {
@@ -25,6 +30,6 @@ public class FreeBusy {
 
 	@Override
 	public String toString() {
-		return "FreeBusy [email=" + email + ", time_slots=" + time_slots + "]";
+		return "FreeBusy [email=" + email + ", time_slots=" + time_slots + ", calendar_id=" + calendar_id + "]";
 	}
 }