+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This section contains changes that have been committed but not yet released.
 ### Fixed
 
 * Fix `Event.hide_participants` not serializing
+* Fix `Event.visibility` not serializing
 
 ### Removed