+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This section contains changes that have been committed but not yet released.
 
 ### Fixed
 
+* Fix `Event.hide_participants` not serializing
+* Fix `Event.visibility` not serializing
+* Fix `FreeBusy` not having a `calendar_id` field
+
 ### Removed
 
 ### Security