+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Nylas Java SDK Changelog
 
+## [Unreleased]
+
+This section contains changes that have been committed but not yet released.
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Fixed
+
+### Removed
+
+### Security
+
 ## [1.20.0] - Released 2023-02-06
 
 ### Added
@@ -348,7 +364,7 @@ This second release aims toward API stability so that we can get to v1.0.0.
 
 Initial preview release
 
-[Unreleased]: https://github.com/nylas/nylas-java/compare/v1.19.2...HEAD
+[Unreleased]: https://github.com/nylas/nylas-java/compare/v1.20.0...HEAD
 [1.20.0]: https://github.com/nylas/nylas-java/releases/tag/v1.20.0
 [1.19.2]: https://github.com/nylas/nylas-java/releases/tag/v1.19.2
 [1.19.1]: https://github.com/nylas/nylas-java/releases/tag/v1.19.1