+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This section contains changes that have been committed but not yet released.
 ### Fixed
 
 * Fix `Tunnel.onMessage` not emitting individual deltas
+* Change `java-websocket` dependency to `api` configuration
 
 ### Removed